+++ xpilot/src/server/netserver.c
@@ -1234,7 +1234,7 @@ static void LegalizeName(char *string)
 	char ch = *string;
 	if (ch == '\"')
 	    ch = '\'';
-	else if (!isprint((int)ch) || strchr("{[]}:,", ch))
+	else if (!isprint(ch) || strchr("{[]}:,", ch))
 	    ch = 'x';
 	*string++ = ch;
     }
@@ -1244,7 +1244,7 @@ static void LegalizeHost(char *string)
 {
     while ( *string != '\0' ) {
 	char ch = *string;
-	if ( !isalnum((int)ch) && ch != '.' )
+	if ( !isalnum(ch) && ch != '.' )
 	    ch = '.';
 	*string++ = ch;
     }
@@ -3117,7 +3117,7 @@ static int str2num (char **strp, int min, int max)
     char *str = *strp;
     int num = 0;
 
-    while (isdigit((int)*str)) {
+    while (isdigit(*str)) {
 	num *= 10;
 	num += *str++ - '0';
     }
